day 3 part 2
(did not commit solution to part one but only involves changing min function)

Part 1 was easy enough for me and I solved it last night. I ended up
struggling with part 2 a lot here and spending a lot of time debugging.
The thing I ended up forgetting was when I added support for tracking steps
at each position was that for the Down and Left directions I was that I
was iterating through those ranges in the reverse order of they way
you'd "step through" manually. This meant that even though the
intersections I was finding were correct, the step values at them could
be incorrect. Reversing the order of iteration for those directions
fixed the problem.