day 6 part 2
My initial approach of going from COM to YOU or SAN and maintaining the
path that was traversed to get to the endpoints, and then subtracting
the common parts out didn't seem to work. I'd like to investigate why
that is but wanted a solution first.

Instead I re-implemented the hash table containing the graph to be
child->parent instead of parent->children and followed the path in
reverse and then computed the difference between the two paths at that
point.