Day 7 part 1
Depends on itertools for permutations because I didn't feel like
implementing that myself. Utilizes the new set_input functionality of my
intcode::Machine to set the input values on each iteration so that it
doesn't have to be done on stdin. For each permutation of 0..5, we
calculate the output value by running the machine's program and then
take the maximum of those.