Refactor intcode machine for day 7 part 2
Allows for machines to output their output values into a
std::sycn::mpsc::SyncSender and for machines to listen to input over a
std::sync::mpsc::Receiver in addition to the existing String and Stdin
methods. This allows machines to be "piped" together.

Additionally adds the ability to "async_execute" a machine by executing
the program on a thread.

This probably breaks a couple of past days but I will perhaps go back
and refactor those later.